Coverity IDs 164020, 164022, 164025
We don't need to check if dot_grid.data is NULL.

Testing a == 0 && b == 0 after testing a nd b individually can't be
correct. Test for both being zero first.

In the case of an error, we could attempt to free memory that had not
been allocated. Prevent this by checking dot_level_pos is non-NULL.