Coverity ID 94819
The comments on clist_end_page() state that it returns negative for an
error, 0 for success and positive for success with a low memory warning.

In fact it always returns 0.

This means we don't conform to the comments, and discard any errors that
occur.

This commit reformats the code to reduce the number of tests and to
store the error codes, as well as adding the potential to return
something other than 0.