pdfwrite - properly process Octal 134 '\' in pdfmark UTF16 string
Bug #698552 "Some PDF outline UTF-16 text is broken"

The escape processing wasn't catering for an escaped escape '\\' which
led to us writing the escaped escape direclty and corrupted the remainder
of the string. We normaly don't use short escapes but write ocral,
because some versions of Acrobat can't handle the short escapes.

This commit just processes the '\\' properly, as per the other
escapes.