Bug 698559 avoid double alpha application
When we have a softmask present in the graphic state, we
wrap up the fill or stroke with an isolated transparency
group in the interpreter.  After the group is pushed we
should set the opacityalpha and the shapealpha to 1 to
avoid a double application of alpha at the rect-fill level and
the group composition level.