PDF interpreter - clamp Box to MediaBox, even when MediaBox is negative
When applying the various 'Box' values from a PDF file, we need to clamp
them to the MediaBox values, so that we don't try and extend the final
output to be larger than the Media (this appears to be Acotbat's
behaviour).

But we also need to account for the fact that the media (and therefore
the content) may not lie up and right. Its possible for the media to
be defined with negative height or width.

We deal with that by scaling/translating the CTM, but we weren't
taking it into account when calculating the actual media size, which led
to us calculating the wrong size.

This is more complicated, but we only do it once per page, so the
performance penalty is insignificant.