PDF interpreter - fix default /Decode for Lab images
When we get an image with no /Decode array we create a default one for
it by looking at the image colour space.

For Lab we use the /Range array, but unfortunately this only covers the
a and b channels, not the L. However, L always ranges from 0 - 100 so
here we just pull the Range array apart, add in that range, and create
a new array for the /Decode.