Tweak art_pdf_composite_pixel_alpha_8.
art_pdf_composite_pixel_alpha_8 is used to blend src and
destination pixels together with a given blend mode.

In the general case, we have n colour compoents, of which the
first n-s are process ones, and the next s are spots (followed
by alpha).

In some cases, we need to blend the first n-s with the given blend
mode, and the remaining s in Normal mode.

This is currently achieved in the code by making 2 different calls
down to art_pdf_composite_pixel_8 for each section of the pixel
(with different 'n' values).

Unfortunately, due to alpha being assumed to be held as the nth
component, this means we need to do a dance where we stash away
a components value, copy the alpha in, call the routine, then
restore the alpha afterwards.

To avoid this, we modify art_pdf_composite_pixel_alpha_8 to take
another int parameter, specifying the position of the first
component that needs to be blended with the Normal mode.

This enables us to simplify the outer routines that call this so
that we have a single call, rather than 3. It also means we don't
recaulate 'scale' etc twice.

I am not expecting any significant speedups from this alone, but
it simplifies the code so that hopefully we might get some in
later commits.