Further optimisation in pdf14_compose_group.
Avoid memcpy's when unnecessary in some of the subfunctions,
by having a 'dst' pointer that points to the results. If
the results are supposed to be the same as the src, don't
copy the source, just reset the dst pointer.

If the dst pointer is set to NULL, then don't copy the results
at all, as it'll be the current values unchanged.