Tweak pdf14_preserve_backdrop (for speed).
First the current code clears the backdrop, then we copy into it.

In many cases we overwrite exactly the area we just cleared. Spot
this, and avoid it.

We *could* optimise this further in cases where we aren't overwriting
exactly the same region. Wait to see if this is justified.