PDF interpreter - permit 0 size font scaling, render nothing
Bug #698619 "PDF 2.0 tests 'zero sized text'"

The PDF 2.0 specification permits font scale of 0 (!), but nothing
should be rendered....

We can't have a font scale of 0, because it breaks our arithmetic, so
we set it to a tiny value, unfortunately this still means that stroking
text rendering modes printed some small garbage.

After a lot of effort with the PDF interpreter, I concluded that there
was no reasonable way to address this in PostScript, as we didn't
retain the font size anywhere and had no reasonable, non-volatile,
place to store it (it needs to respect gsave and grestore).

So in the end I added it to the graphics state and created two new
operators to read and write it (and undefined them after startup).
Its an unfortunately brute-force approach, but it has the benefit of
actually working.