PDF Interpreter - support PDF 2.0 SpotFunction arrays
Bug #698620 "PDF 2.0 tests invalid halftones"

This was an oversight on my part, I missed new functionality when
reading the PDF 2.0 specification.

The SpotFunction in a type 1 Halftone can now be a name, function or (new)
an array of names. If its an array we are supposed to check the entries
in turn selecting the first one we recognise. If we don't recognise any
then we should use the default.