PDF and PS Interprters and ps2write - fix type 6 halftones & transfers
Bug698621 "PDF 2.0 tests not applying transfer function from halftone"

The halftones in question are defined (by the PDF interpreter) as type
6 halftones, ie file-based threshold arrays.

Rather to my surprise, the reason this didn't work turns out to be
because the PostScript interpreter has never supported transfer functions
with type 6 halftones.

So; add support for transfer functions in type 6 halftones. This exposed
a problem with ps2write (and potentially pdfwrite) when emitting these
transfer functions, instead of simply specifying the object number of
the function, we were writing it out decorated with "/TransferFunction"
which caused ps2write output to fail. So fix that here too.

Remove the 'OBSELETE' comment from the halftone structure, I don't know
who thought this was obselete, but it really does not seem to be.