Documentation - remove references to .runandhide
We removed a number of PostScript operators from accessibility after
the interpreter is running as part of the security overhaul. One of
these was .runandhide but we accidentally left a couple of references
in the documentation (the main part of the documentation was correctly
updated).

So remove the references here, and update the documentation to note the
potential risks in using -dNOSAFER with a save object on the exec
stack.