Bug 698647: pdfwrite: avoid cached glyphs for Type 3 font input only
The original fix in 8360852efab5643d93cc3b040832075e199cd205 wrongly assumed
that the changed code only came into effect for Type 3 font *input* where,
in reality, it's for any font for which we have to create a Type 3 in the
output.

The change to disable the use of the cache should only apply to Type 3 input
where we want to force the execution of the BuildGlyph/BuildChar proc so
pdfwrite can capture it.

In all other cases, we need to leave the cache in play.