Bug 698672: Fix new scanconverter/alphabits interaction.
The alphabits mechanism relies on the scan converter not
writing trapezoids so tall as to cause its line buffers
to flush too early.

This is achieved by honoring the max_band_height field in
the device header. Previously the new scan converter was
ignoring this. Here we alter the code to pay attention to
it.