Remove useless test in PDF interpreter transparency code
This test was always true since we always have .begintransparencygroup
It was changed in 2008 from a check for .inittransparencymask which was
created back in 2006. It was probably a test in case the PDF interpreter
was running on a system that may not include the pdf14 compositor, but
that is only a guess.