Fix Bug696372: Transparency with DeviceN not coping with bad PDF's
When a PDF with transparency stops interpretation due to an error, the
grestore may have popped the pgs so that the pdf14 compositor is no
longer the currentdevice. The logic in pad14_spot_get_color_comp_index
assumed that the device was a pdf14_device and got a bogus pointer to
devn_params. Instead, use the device's ret_devn_params function to get
the actual pointer.

Also, the poppdf14devicefilter would execute without the pdf14 device
as the currentdevice. Save the annots_gstate explicitly in pdf_main
showpagecontents so it can be set back before doing annots and the
poppdf14devicefilter.

This also required gs_setgstate to always do gs_do_set_overprint since
the setgstate may have skipped over states with the overprint compositor.
This caused "drawn_comps" to not be set correctly for the pdf14
compose_group.

Fixups for 696372 fix commit (also ne->be last line of commit msg)