PDF interpreter - try and catch yet more cases of circular references
No bug report for this one, the customer requested destruction of the
PDF file after analysis, and there was no way to reduce the file to the
point where nothing sensitive remained.

The problem is that a Form XObject referenced a number of Image XObjects
one of which contained a DecodeParms array, where the second element
referenced the original Form XObject. This circular reference leads to
an error.

Here we maintain a dictionary (initially empty) on the stack which we
populate with object numbers as we recursively dereference objects.
When we start a new composite object we copy the array into a new array
which we discard on completion. This is to prevent false detection of
circular references if we dereference multiple composite objects at the
same level (eg if we deal with multiple images, each of which perhaps
contained a reference to the same colour space).

We now issue an error quoting the object number being circularly
referenced, replace the object with a null object and continue.

No differences expected.