Avoid clearing cm_comps arrays in cmap functions.
This goes against commit 99c48263b9 by Ray (dating from Nov 1 2009)
to fix Bug 690713 and solve some indeterminisms.

My belief is that this SHOULD be fixed in the devices, not in the
callers. Accordingly I have backed that out, and changed the
cmyk_cs_to_psdcmyk_cm function (the only one I can find that does
not set everything to be 0).

The cluster now shows this running with no diffs.