Tweak map_XXXX_subclass functions.
To cope with subclassed devices, we no longer simply call:

 pprocs = dev_proc(dev, get_color_mapping_procs);
 pprocs->map_cmyk(dev, ...)

etc. The 'done thing' is to run up the the dev->parent pointers
as far as we can, and to get the color mapping procedures from
that. These color mapping procedures then need to be called
using the dev pointer from which they were fetched.

To do this, we provide get_color_mapping_procs_subclass
and map_XXX_subclass functions that pickle the dev search.

The only downside to this is that we end up doing the search
up the tree twice (or more than twice in the case where we
make several mapping calls).

Here we therefore tweak these functions so that
get_color_mapping_procs_subclass returns both the procs AND
the dev to use when calling them in a structure, and the
map_XXX_subclass functions now take that structure.

Broadly, this change shouldn't actually alter any operation,
other than being slightly more efficient.

There are a few wrinkles:

1) in gsicc_replacecm.c, we were fetching the procs using the
dev chasing function, but calling them using the unchased dev
value. Fixed here to use the matching dev value in both cases.

2) In a couple of places, we have special handling for forwarding
devices. I am not convinced that we handle subclassed forwarding
devices correctly (or devices that forward to subclassed devices).
I have marked these areas with FIXMEs, but they are no worse
now than they were before.