New devices - pdfimage and PCLm
These devices render input to a bitmap (or in the case of PCLm
multiple bitmaps) then wraps the bitmap(s) up as the content of
a PDF file. For PCLm there are some additional rules regarding
headers, extra content and the order in which the content is
written in the PDF file.

The aim is to support the PCLm mobile printing standard, and
to permit production of PDF files from input where the graphics
model differs significantly from PDF (eg PCL and RasterOPs).

Devices are named pdfimage8, pdfimage24, pdfimage32 and PCLm.

Currently produce valid PDF files with a colour depth of 8 (Gray),
24 (RGB) or 32 (CMYK), the PCLm device only supports 24-bit RGB.

Devices supports the DownScaleFactor switch to implement
page level anti-aliasing

-sCompression can be set to None, LZW, Flate, JPEG or RLE (LZW
is not supported on PCLm, None is only available on PCLm for
debugging purposes).

The PCLm device supports -dStripHeight to set the vertical height
of the strips of image content, as required by the specification.

For JPEG compression the devices support both the JPEGQ and
QFactor controls, exactly as per the jpeg and jpeggray devices.