Fix bug 695877: SEGV when display callback not set.
Change most of the device procs to return gs_error_Fatal when the callback
hasn't been set. The open is allowed to be called before it is set, but any
operations other than fill_rect (allowed for due to fillpage) and close will
now return Fatal errors. This allows the device to be "fake opened" by gs_init
before the parameters are processed and the display callback is set and the
device is opened for actual use.