Fix bug 696845 SEGV with --saved-pages-test and transparency
When a file pushes a pdf14devicefilter (or other compositor) it will
still be the currentdevice after the filter is popped when the saved
pages printing occurs. We need to check if the device is a forwarding
device (as compositors will be when deactivated) and use the target
device to print the saved pages.

Also, if a device uses the gx_default_dev_spec_op, it will return 0
from the supports_saved_pages call even when it is a printer device.
In gdev_prn_forwarding_dev_spec_op, check for a zero return and return
true (1). A device that is a printer device that doesn't support saved
page printing needs to return < 0 to prevent this.

Add a finalize method for gx_device_printer so we can free up the saved-
pages-list when the device is freed, but not when the device is closed
which can happen if it needs to close due to put_params. Also the
return code from the output_page in gx_saved_page_params_process was
ignored. Both seen with the file Bug687111.ps and the psdcmyk device.

Also for devn devices, such as psdcmyk, we need to save the separation
(spot color) names collected during the execution of setcolorspace in
the gx_saved_page and set them in the devn_params when rendering. These
are not handled by the paramlist.

Disable --saved-pages= and --saved-pages-test with PCL and XPS since it
was never completely implemented and the partial implementation causes
many errors when tested with --saved-pages-test. Note that while the
--saved-pages-test option is still accepted as a parameter, it is ignored
so that regression testing with --saved-pages-test can run.

Regression shows 3343 "diffs", but bmpcmp doesn't show any diffs.