Guard against divide-by-zero exceptions in the halftone code
Bug #697663 "FPE in gx_compute_cell_values()"
Bug #698332 "FPE (Division by zero) in pick_cell_size()"

These have the same root cause, in pick_cell_size() we can end up with
halftone cell parameters which are (I think) improperly initialised.

Certainly its possible to get to the point where we divide using one
of the parameters, and if its 0 we get an error. I suspect that the
halftone may be badly broken even when its not 0, but both these files
are the result of fuzzing, so I doubt its possible to have a real
working file get to this point.

Guard against the divide-by-zero errors by checking that both the
parameters are non-zero (its permissible for one to be 0 I believe)