Bug 698733: make vector device memory stable
The pcl/xl device (gdevpx.c) was using "normal" Postscript global memory for
its own use, and the vector base device on which it is based. That means the
memory it allocated could disappear with the end-of-job restore in the
Postscript interpreter.

Change it so it uses stable memory which is still garbage collected, and still
subject to relocation, but is *not* subject to Postscript save/restore
operation.