Repair the device API - device methods should never be NULL
The original intention of the device API was that device methods (with
the odd exception of fill_rectangle) must not be NULL. That appears
to have been broken with the introduction of the transparency
compositor, with the result that the code now:

1) has to check for methods being NULL sometimes, because that causes
crashes.

2) Uses a device method being NULL to determine code paths.

We want to return the device method design, because having to check for
NULL methods before calling is wasteful.

In this commit:

define default methods for all the methods that previously did not have
one.

Modify gx_device_fill_in_procs to include those defaults in the device
methods which it checks for NULL and replaces.

Finding all the places where the code path differs if a device method
is NULL, alter to check against the default method instead.

Remove NULL checks where they are no longer required. Device whose
'procs' get copied directly (clist) must have defaults for all methods.

Fix gx_copy_device_procs so that we only special case (don't replace
the method) for those few methods where it matters that the subclassing
device must have the default method if the subclassed device does. Long
term we should change the code not to rely on this!

Remove a couple of macros that were not required or plain dumb.