pdfwrite - fix an error handling problem
Bug 698735 "Ghostscript: SEGV in pdfwrite with -Z@"

This was, unfortunately, caused by the shift to cleaning up memory
instead of relying on the garbage collector.

When an error occurred writing the metadata, the code simply exited
back to the caller, it did not close the temporary stream it was
writing the metadata to.

Due to the way that 'asides' are handled this led to the wrong stream
being closed, and a pointer retained to a closed stream. The result
was the seg fault.

There may well be other conditions like this still present.