Avoid buffer overflow in GraphicsAlphaBits cdoe with planar device.
In GraphicsAlphaBits, we get the pixels from the underlying device
(in chunky format). The buffer we get that data into was sized
according to get_device_raster(dev, false), which (for a planar
device) returns the size for line of planar data. We want the
size for one line of chunky data, so use get_device_raster_chunky
instead.