Fix operation of 1bpc, 3 component halftones.
In the chunky world, it makes no sense to deal with 3 bit rgb as
this doesn't pack neatly into a byte. You'd do it by handling
4 bit chunks.

Accordingly, parts of Ghostscript are written to assume that you'd
not do anything this crazy.

Unfortunately, in the planar world, a 1 bpc, 3 component device
does make (some) sense. Sadly, stuff like the halftone generation
gets deeply confused by this.

Here we fix 2 specific instances of this problem. In the halftone
generation, we generate 3bit chunky halftones as 4 bit ones, and in
the planar_copy_color routines, we assume that 3 bit data is
actually 4 bit (as a) it probably comes from these routines, and
b) if it doesn't it'd be packed as 4 bit in any sane world).