ps2write - fix string splitting for output formatting
Bug #698708 "Ghostscript loops"

When writing output for ps2write, we limit the output width to 255
characters. If a string is longer than that we try to split it on
PostScript tokens. If we can't do that then we just break at 255
characters and stick in a newline.

The logic in here however was flawed however, it failed to update the
pointer 'p' which points to the character currently being processed
and did not reset the width count to 0 when the string was output.

This could lead to us in an infinite loop.