Fix buffer overrun in gp_enumerate_files_next.
If pattern="*" and patlength = 1, then pathead = 1.

When we look for any following subdirectory, the code attempts
to look for the next subdirectory name. To do this it starts
searching from the byte after the position of the current '/'.
Unfortunately, if the string ends at the current wildcard, we
don't have a '/', we have a NUL terminator. This means the
current code starts looking just after that.

In code terms this means the current code accesses at
pattern + pathead + 1, without checking pattern[patthead] first.
This is a simple fix.

We also take this opportunity to simplify a couple of while loop
conditions. The extra negation is too much for my tiny brain to
cope with.