Tweak the gs_register_root API
The existing gs_register_root() API allows for two ways to call the function.

The first is for the client to pass in a pointer to a garbager root structure
to use, allowing the client to hold onto that root so it can, later,
unregister the root.

The second, the client passed in a NULL pointer for the garbager root structure,
and the allocator will allocate its root internally. Roots such as these are
not visible/accessible outside the allocator.

The problem with the former is that it requires the full definition of the
gs_gc_root_t type to be available to the caller: that is a dependency nightmare
(due to assumptions in the original design).

The problem with the second is that the caller has no record with which to
unregister the root.

This commit changes the API so it takes a pointer to a pointer to gs_gc_root_t.

If the gs_gc_root_t ** parameter is NULL, the behaviour is the same as the
existing code (the root is allocated and only visible inside the allocator).

If the gs_gc_root_t ** points to a "real" gs_gc_root_t *, that is the client
supplied root.

The new part is if gs_gc_root_t ** points to a NULL gs_gc_root_t *, the
allocator will allocate a new root, and return it to the client via the
pointer-pointer parameter.

Importantly, this allows gs_gc_root_t to be opaque for client code.