PDF interpreter - move a default value to be compatible with all xref types
The commit df5b3426d31f79c13a735dff9118e9798ce97af9 which works around
PDF files where the /N value and the actual ICC profile colour space
number of components do not match.

Unfortunately, I made a minor error when creating this, which wasn't
exposed by the test files on hand at the time. I set the default value
of ICCProfileNError to false when processing an xref of type 'orig'
(ie not an xref stream) which means that for PDF files which use xref
streams the value was not defined, leading to an error.

Move the initial declaration to pdfopenfile instead so that it works
no matter what type of xref we have,.