Ensure that single component planar devices can cope with hl_color.
As part of tracking down indeterminisms in the cluster, I found that:

 gs -sBandListStorage=file -sOutputFile=out%d.psd -dMaxBitmap=10000
 -sDEVICE=psdcmyk -r300 -Z: -sDEFAULTPAPERSIZE=letter -dNOPAUSE
 -dBATCH -K2000000 -dClusterJob -dJOBSERVER -ZL
 tests_private/pdf/sumatra/586_-_missing_images_gs_SMask_not_applied.pdf

was hitting a case where we were writing a pattern tile into the
clist that hadn't been initialised at all. This was because all
the writes to the pattern accumulator went through
pattern_accum_fill_rectangle_hl_color. This passes the hl_color
data to the memory devices fill_rectangle_hl_color function, and
fills the mask with 1s.

The mask was being entirely filled with 1s and so was being discarded
as unnecessary. Unfortunately the devices fill_rectangle_hl_color
entrypoint was left as the default which does nothing by raise an
error.

The fix here is to ensure that when we open a planar device, we
always add in a suitable fill_rectangle_hl_color function, even
when the device is superficially the same as a chunky one.

Thanks are due to Michael and Ray for their help with tracking this
down.