Fix indeterminism caused by buffer overrun in thresholding.
Running:

 gs -sOutputFile=out%d.pbm -dMaxBitmap=400000000 -sDEVICE=pbmraw -r300
 -sDEFAULTPAPERSIZE=letter -dNOPAUSE -dBATCH -K2000000 -dClusterJob
 -dJOBSERVER /home/marcos/cluster/tests_private/pdf/uploads/Bug698519.pdf

then catting out{1,2,3,4,5}.pbm together and md5summing it gives a
different result to running the above job direct to out.pbm.

After much nightmarish debugging, it transpires that this is because
the special case in the SSE thresholding code that spots the 200%
scale case is being triggered incorrectly.

The image is 1753 pixels wide, which would be 3506 pixels at 200%.
We actually display it at 3507 pixels. The code was triggering the
200% case by calculating a scale_factor value of fixed_half (128).

The rounding here was causing a false positive.

Instead, we skip the calculation and check by directly comparing
source and dest sizes.