Bug 698773: Cope with "charstring" of invalid type
This file has a font whose CharStrings dictionary contains:
/space 3 def

Ideally, when we hit such a case, we should fallback to the .notdef glyph,
unfortunately, it also has:
/.notdef 0 def

We *should* throw an error since this is clearly totally invalid, but other
interpreters *appear* to simply ignore the problem. So, we opt to return an
empty charstring.