OpenType fonts - regard fonts with multiple GSUB tables as invalid
Bug #697627 "IO in gs_notify_all function"

The bug title is a misnomer; the actual problem is that we have a badly
corrupted font which appears to have 2 GSUB tables. This causes us to
allocate two chunks of memory to hold the tables, losing the original
reference when we allocate the second, and also to register the font
to have the GSUB table released twice.

Trying to release the same memory twice leads to us trying to access
invalid memory, and therefore to a seg fault.

So if we find we have a GSUB, and we've already allocated memory for a
GSUB table, just throw an error.

Chris thinks it may be possible to legally have multiple GSUB tables, if
we ever find such a thing we'll have to rethink this.