Bug 698784: Fix the alias for Helvetica-Narrow-Bold-Oblique
Init/Fontmap.GS:

There was a misalignment between the filename of the
NimbusSansNarrow-BdOblique and its /Fontname in the T1 font itself.

Filename:   NimbusSansNarrow-BdOblique
/Fontname: /NimbusSansNarrow-BoldOblique

This worked correctly if the fonts were located directly in
/usr/share/ghostcript/Resources/Font folder on the filesystem.

However, on Fedora we are using /usr/share/fonts/urw-base35/ folder,
which is part of Ghostscript's Search Path. In this case it was
causing the Ghostscript being unable to locate the correct font, thus
crashing while opening any document containing Helvetica Narrow Bold
Oblique font...

*In addition*, rename the font file so it matches.