Bug 698795: --without-versioned-path option introduced
Using this option will result in Ghostscript being installed into a
paths which do not contain its version in them, which can be useful
on some distributions.

As a result of using this option, the Ghostscript's search path will
be updated as well, to include correct files locations.

This option is disabled by default, and configure's help page states
that using this option is dangerous, risky and unsupported.