pdfwrite - colour conversion changes
Bug #698723 "convert rgb pdf to cmyk with icc profile yields wrong black"

When converting to a base space colour, and applying the transfer
function, don't use the old-fashioned colour conversion routines but
instead use the ICC profile colour management.

With this change in place, the CMYK components in the output file are
the same as those when rendering to TIFF CMYK output.

I am convinced there is more work to do in this area.