Bug 698814: use --docdir= configure parameter
Previously, the Ghostscript was using custom path to the documentation
(as docdir=$(gsdatadir)/doc). This was causing that value of --docdir=
parameter of ./configure was accepted, but not used at all.

This commit fixes this issue, by using docdir=@docdir@@VERSIONED_PATH@
instead. However, as a side effect this results in default path for
documentation to changed to this (with --prefix=/usr):

  /usr/share/doc/ghostscript/<version>/

To stay backward compatible, a symlink is automatically created to
point from the old location (/usr/share/ghostscript/<version>/doc) to
the new location.

NOTE: Trying to fix this issue with the location of documentation
      staying as it was would require some hacking inside the Autoconf
      itself, which is not desirable.