Segv fix for XPS files with trans out to sep device
When going out to a separation device, PCL and XPS should set
PageSpotColors to 0.  For PCL this lets the sep device know
that there are not any spot colors on the page.

For XPS the situation is a bit more complex.  XPS files with named colors
are very rare.  I have never seen one in the wild.  Supporting
the separation of them will require work in the XPS interpreter
to keep track of the colorants (the names of which are in the colorant
tags of the ICC profile) and make sure that the values are properly
mapped to the colorant positions on the output device.  We will also
want to do a pre-parser to count the number of independent colorants
on the page to ensure proper set up of the device.  This is needed
especially if the page has transparency as the pdf14 device expects to
have this information.   Again running into them in the wild is not going to
occur.  If for some reason it does become an issue in the future,
we can invest the time in doing the above.  In the meantime if they
are encountered they will be remapped to equivalent CMYK colors instead
of causing a segv.